refactor model package to dto
