+++ backend/src/main/java/com/llm_service/llm_service/controller/conversation/ConversationController.java
@@ -1,7 +1,7 @@
 package com.llm_service.llm_service.controller.conversation;
 
+import com.llm_service.llm_service.dto.Conversation;
 import com.llm_service.llm_service.exception.conversation.ConversationNotFoundException;
-import com.llm_service.llm_service.model.Conversation;
 import com.llm_service.llm_service.service.ConversationService;
 import java.util.List;
 import java.util.UUID;
