+++ backend/src/main/java/com/llm_service/llm_service/controller/conversation/ConversationControllerMapper.java
@@ -1,6 +1,6 @@
 package com.llm_service.llm_service.controller.conversation;
 
-import com.llm_service.llm_service.model.Conversation;
+import com.llm_service.llm_service.dto.Conversation;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
