+++ backend/src/main/java/com/llm_service/llm_service/controller/discussion/DiscussionControllerMapper.java
@@ -1,7 +1,7 @@
 package com.llm_service.llm_service.controller.discussion;
 
-import com.llm_service.llm_service.model.Conversation;
-import com.llm_service.llm_service.model.Discussion;
+import com.llm_service.llm_service.dto.Conversation;
+import com.llm_service.llm_service.dto.Discussion;
 import java.util.UUID;
 import org.mapstruct.Mapper;
